requirements.txt
torch
fastapi
fastapi-cli
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
pydantic
pydantic-settings
prometheus-client
scikit-learn
transformers
torch
numpy
pandas
mlflow
dvc[s3]
boto3
nltk
spacy
pytest
pytest-cov
pytest-asyncio
python-dotenv
pyyaml
cryptography
seaborn